Server Complete, Postman missing 1 VK test

+++ finalcheckpoint-server/Services/VaultKeepsService.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using finalcheckpoint_server.Exceptions;
 using finalcheckpoint_server.Models;
 using finalcheckpoint_server.Repositories;
 using Microsoft.AspNetCore.Mvc;
@@ -10,15 +11,22 @@ namespace finalcheckpoint_server.Services
     {
         private readonly VaultKeepsRepository _repo;
 
-        public VaultKeepsService(VaultKeepsRepository repo)
+        private readonly VaultsRepository _vrepo;
+
+        public VaultKeepsService(VaultKeepsRepository repo, VaultsRepository vrepo)
         {
             _repo = repo;
+            _vrepo = vrepo;
         }
 
-        // internal IEnumerable<VaultKeep> Get()
-        // {
-        //     return _repo.Get();
-        // }
+        internal IEnumerable<VaultKeep> GetAll()
+        {
+
+            IEnumerable<VaultKeep> vaultkeeps = _repo.GetAll();
+            return vaultkeeps;
+        }
+
+
 
         internal VaultKeep Create(VaultKeep newVau)
         {
@@ -27,8 +35,11 @@ namespace finalcheckpoint_server.Services
             return newVau;
         }
 
-        public string Delete(int id)
+        internal string Delete(int id, string userId)
         {
+            Vault original = _vrepo.GetById(id);
+            if (original == null) { throw new Exception("Bad Id"); }
+            if (original.CreatorId != userId) { throw new Forbidden("Access Denied: You are not the original creator"); }
             _repo.Delete(id);
             return "Delete Successful";
         }
